memory_patcher: Add ability to detect patched memory
Adapt the memory patching code to add detection logic so we can detect when
memory has already been patched. Look for this condition in patcher_open()
and relay the return code OPAL_ERR_RESOURCE_BUSY when the memory has been
previously patched. If we detect this, also unapply any patches that we've
applied.

Signed-off-by: Rich Welch <rlwelch@amazon.com>